+++ fltk-sys/cfltk/cfl.h
@@ -56,6 +56,8 @@ void Fl_paste(Fl_Widget *, int src);
 
 void Fl_set_scheme(const char *scheme);
 
+int Fl_scheme(void);
+
 unsigned int Fl_get_rgb_color(unsigned char r, unsigned char g, unsigned char b);
 
 const char *Fl_get_font(int idx);